// src/types/productTypes.ts
export interface Market {
  market: string
  division: string
  min: number
  max: number
}

export interface Change {
  dir: 'up' | 'down'
  pct: number
}

export interface Product {
  id: number
  slug: string
  nameBn: string
  category: string
  categoryNameBn: string
  categoryIcon: string
  unit: string
  image: string
  today: number
  yesterday: number
  lastWeek: number
  lastMonth: number
  change: Change
  markets: Market[]
}